feat: protocol update to 13.16 (#1010)
This changes only includes the protocol update up to 13.16. For the changes related to the features, please check PR #531 
Disclaimer: this changes removes tournament system and coins completely since Cipsoft removed it from their client, no way back.

Client for login: https://github.com/dudantas/tibia-client/releases/tag/13.16.13209
Client notes:
• Use the client "client-127.0.0.1-13.16.13209" to test local, it already uses the OpenTibia RSA key and the local ip.
• Place the files "cache/eventscheduler.json" and "cache/boostedcreature.json" as read-only for the correct functioning of the systems by the server.

+++ src/lua/functions/creatures/combat/spell_functions.cpp
@@ -26,7 +26,7 @@ int SpellFunctions::luaSpellCreate(lua_State* L) {
 	SpellType_t spellType = SPELL_UNDEFINED;
 
 	if (isNumber(L, 2)) {
-		int32_t id = getNumber<int32_t>(L, 2);
+		uint16_t id = getNumber<uint16_t>(L, 2);
 		RuneSpell* rune = g_spells().getRuneSpell(id);
 
 		if (rune) {
@@ -191,7 +191,7 @@ int SpellFunctions::luaSpellId(lua_State* L) {
 		if (lua_gettop(L) == 1) {
 			lua_pushnumber(L, spell->getId());
 		} else {
-			spell->setId(getNumber<uint8_t>(L, 2));
+			spell->setId(getNumber<uint16_t>(L, 2));
 			pushBoolean(L, true);
 		}
 	} else {
